ang-crud3 terminado

+++ ang-cruds/ang-crud3-firestore/src/app/components/products/products.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ProductService } from '../../services/product.service';
 import { Product } from 'src/app/models/product';
+import { config } from 'rxjs';
 
 @Component({
   selector: 'app-products',
@@ -10,6 +11,10 @@ import { Product } from 'src/app/models/product';
 export class ProductsComponent implements OnInit {
 
   products = [];
+
+  editingProduct: Product;
+
+  editing: boolean = false;
 
   constructor(private productService: ProductService) { }
 
@@ -21,7 +26,20 @@ export class ProductsComponent implements OnInit {
   }
 
   deleteProduct(event, product) {
-    this.productService.deleteProduct(product);
+    if(confirm('Are you sure?')){
+      this.productService.deleteProduct(product);
+    }
+    
   }
 
+  editProduct(event, product: Product) {
+    this.editingProduct = product;
+    this.editing = !this.editing;
+  }
+  
+  updateProduct() {
+    this.productService.updateProduct(this.editingProduct);
+    this.editingProduct = {} as Product;
+    this.editing = false;
+  }
 }
